+++ exhibitera/changelog.md
@@ -15,6 +15,7 @@ _Exhibitera Hub and Apps must all be upgraded to version 6 at the same time._
 
 ### Media Player
 - Adjust the opacity of watermarks
+- Annotate media with custom text
 
 ### Word Cloud
 - Switched profanity filtering library, so please check your word clouds